Moved incremental smt struct encoding to subfolder
As support for non-native C types (e.g. enums) is added in the
incremental smt2 solver, other new encoding functions will be added. For
this reason, to avoid putting all the encoding functions in the
incremental smt2 solver root folder a new `encoding` folder has been
created and the struct encoding functions have been moved there.
