+++ src/solvers/smt2_incremental/encoding/struct_encoding.h
