+++ unit/solvers/smt2_incremental/encoding/module_dependencies.txt
@@ -1,0 +1,3 @@
+solvers/smt2_incremental/encoding
+testing-utils
+util
